+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import MainContent from "@/contents/main";
 import { getAllExperiments } from "@/lib/helpers/experiments";
 import { getTemplatesFromCDN } from "@/lib/helpers/templates";
 
-export const revalidate = 3600;
+export const revalidate = 86400;
 
 export default async function Home() {
   const templates = await getTemplatesFromCDN()