+++ src/app/templates/page.tsx
@@ -2,21 +2,42 @@ import { Metadata } from "next";
 import TemplatesMainContent from "../../contents/templates";
 import { getTemplatesFromCDN } from "@/lib/helpers/templates";
 import { TEMPLATE_FILTER_NAMES, TemplateFilterName } from "@/lib/constants/filters";
+import { absoluteURL, createMetaAlternates } from "@/lib/utils";
 
-export const metadata: Metadata = {
-     title: "Շաբլոններ"
-}
-
-export const revalidate = 3600;
-
-export default async function TemplatesMainPage({ searchParams }: {
+interface PageProps {
      searchParams: Promise<{
           page?: string;
           pageSize?: string;
           query?: string,
           category?: TemplateFilterName
      }>;
-}){
+}
+
+export const generateMetadata = async({searchParams}: PageProps): Promise<Metadata> => {
+     const search = await searchParams
+     const result = await getTemplatesFromCDN();
+     const allowedPageSizes = [4, 8, 16, 24];
+     const requestedPageSize = Number(search.pageSize);
+     const pageSize = allowedPageSizes.includes(requestedPageSize) ? requestedPageSize : 8;
+     const query = search.query?.trim().toLowerCase() ?? ""
+     const searchResults = result.filter(item =>!query || item.title.includes(query))
+     const totalResults = searchResults.length
+     const totalPages = Math.max(1, Math.ceil(totalResults / pageSize))
+     const currPage = Math.min(Math.max(Number(search.page) || 1, 1),totalPages)
+     return {
+          title: "Շաբլոններ",
+          description: "Հատուկ շաբլոններ՝ գիտափորձեր կատարելու համար",
+          pagination: {
+               previous: currPage > 1 ? absoluteURL(`/templates?page=${currPage - 1}`) : undefined,
+               next: currPage < totalPages ? absoluteURL(`/templates?page=${currPage + 1}`) : undefined
+          },
+          alternates: createMetaAlternates("/templates")
+     }
+}
+
+export const revalidate = 86400;
+
+export default async function TemplatesMainPage({ searchParams }: PageProps){
      const params = await searchParams;
      const result = await getTemplatesFromCDN();
      const allowedPageSizes = [4, 8, 16, 24];