+++ src/lib/helpers/templates.ts
@@ -1,7 +1,8 @@
 import type { Template } from "@/lib/types";
 import { absoluteCDN } from "../utils"
+import { cache } from "react";
 
-export const getTemplatesFromCDN = async (): Promise<Template[]> => {
+export const getTemplatesFromCDN = cache(async (): Promise<Template[]> => {
      try {
           const res = await fetch(absoluteCDN("/templates.json"));
           if(!res.ok) return []
@@ -10,4 +11,4 @@ export const getTemplatesFromCDN = async (): Promise<Template[]> => {
      } catch {
           return []
      }
-}
+})